fix: add mocks to a separate file.
