+++ backend/product_service/lambda/mocks.ts
@@ -1,0 +1,23 @@
+export const products = [
+  {
+    id: 'product_1',
+    title: 'Product 1',
+    description: 'This is the first product',
+    count: 10,
+    price: 100,
+  },
+  {
+    id: 'product_2',
+    title: 'Product 2',
+    description: 'This is the second product',
+    count: 20,
+    price: 200,
+  },
+  {
+    id: 'product_3',
+    title: 'Product 3',
+    description: 'This is the third product',
+    count: 30,
+    price: 300,
+  },
+];
